Just did: Waits for data of all the users to come in before it sets the data.
Issues:
the length of the data needs to automatically calculated as this rn is hardcoded.

Recent Update:
Feed Screen is able to store and process video data from firebase and integrate this into an  object which the data can be fetched from.

To do:
Timing Out issues flatList is processed before data is fetched