Just did: Slight progress on user making comment
issues: it fails when it tries to make a comment, the system thinks the string is empty, this probably that to do with the fact that render is prefected. Might need to refactor code in bottom sheet

To do: User should be able to make a comment on the page increasing the comment Count.
Change the layout styling as I don't like the way it looks.