Just did: Fixed typo in query script stopping the comments from loading Fixed navigator screen width taking the full size of the screen.
To do:
Change the comment page styling, set conditions allowing the user to make a comment if the comment page is currently empty.
Keep track of comment count when a user adds a comment.

Previous work: Users can now make a comment on a post

Refactored comments code,so it's easier to pass parameters between objects